feat: move excel to common
