+++ packages/common/excel/src/domain/column-parameters.ts
