+++ packages/common/excel/src/domain/excel-builder-text-column-interceptor.ts
