+++ packages/common/excel/src/domain/export-excel-by-entries.ts
