+++ packages/common/excel/src/domain/prepared-excel-workbook.ts
