+++ packages/common/excel/src/domain/width-of-symbols-map.json
