+++ packages/common/integration/apps/excel/assets/favicon.ico
