+++ packages/common/integration/apps/excel/polyfills.ts
