+++ packages/common/integration/apps/excel/tsconfig.app.json
@@ -1,0 +1,9 @@
+{
+    "extends": "../../../tsconfig.base.json",
+    "compilerOptions": {
+        "paths": {
+            "@angular-ru/common/*": ["../common/*/public_api.ts"]
+        }
+    },
+    "include": ["main.ts", "polyfills.ts"]
+}
