+++ packages/common/integration/tests/excel/helpers/file-suites/test-4-options.xls
