+++ packages/common/integration/tests/excel/helpers/file-suites/test-6-auto-width.xls
