+++ packages/common/integration/tests/excel/helpers/file-suites/test-7-auto-width-translate.xls
