+++ packages/common/integration/tests/excel/helpers/file-suites/test-8-auto-width.xls
