--- packages/ng-excel-builder/ng-package.json
@@ -1,13 +1,0 @@
-{
-    "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-    "lib": {
-        "entryFile": "./src/public_api.ts",
-        "umdModuleIds": {
-            "@angular-ru/common/typings": "angular-ru-common-typings",
-            "@angular-ru/common/webworker": "angular-ru-common-webworker",
-            "@angular-ru/common/date": "angular-ru-common-date",
-            "@ngx-translate/core": "ngx-translate-core"
-        }
-    },
-    "dest": "./dist/library"
-}
