--- packages/ng-excel-builder/package.json
@@ -1,22 +1,0 @@
-{
-    "name": "@angular-ru/ng-excel-builder",
-    "version": "15.307.0",
-    "description": "Angular-RU package",
-    "homepage": "https://github.com/Angular-RU/angular-ru-sdk/tree/master/packages/ng-excel-builder#readme",
-    "bugs": "https://github.com/Angular-RU/angular-ru-sdk/issues",
-    "repository": "git+https://github.com/Angular-RU/angular-ru-sdk.git",
-    "license": "MIT",
-    "author": "angular-ru@yandex.ru",
-    "scripts": {
-        "build:app": "ng build ng-excel-builder-app",
-        "build:lib": "ng build ng-excel-builder-library && cp README.md dist/library",
-        "deploy": "echo 1; #cp README.md dist/integration && ng deploy ng-excel-builder-app",
-        "start": "ng serve ng-excel-builder-app",
-        "test": "yarn jest --config ./jest.config.js",
-        "postversion": "npe version \"$(npe version)\" --package=\"$PWD/dist/library/package.json\""
-    },
-    "publishConfig": {
-        "access": "public",
-        "directory": "dist/library"
-    }
-}
