--- packages/ng-excel-builder/tsconfig.app.json
@@ -1,11 +1,0 @@
-{
-    "extends": "./tsconfig.base.json",
-    "compilerOptions": {
-        "paths": {
-            "@angular-ru/ng-excel-builder": ["./src/public_api.ts"],
-            "@angular-ru/common/*": ["../common/dist/library/*"]
-        }
-    },
-    "files": ["integration/app/main.ts", "integration/app/polyfills.ts"],
-    "exclude": ["integration/tests"]
-}
