--- packages/ng-excel-builder/tsconfig.base.json
@@ -1,12 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "rootDir": ".",
-        "baseUrl": ".",
-        "typeRoots": ["../../node_modules/@types", "./"],
-        "paths": {
-            "@angular-ru/common/*": ["../common/*/src/public_api.ts"],
-            "@angular-ru/ng-excel-builder": ["./src/public_api.ts"]
-        }
-    }
-}
