--- packages/ng-excel-builder/tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "extends": "./tsconfig.base.json",
-    "references": [
-        {
-            "path": "./tsconfig.lib.json"
-        },
-        {
-            "path": "./tsconfig.app.json"
-        },
-        {
-            "path": "./tsconfig.spec.json"
-        }
-    ]
-}
